% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.survival-package}
\alias{mlr3learners.survival}
\alias{mlr3learners.survival-package}
\title{mlr3learners.survival: Learners from the {survival} package for 'mlr3'}
\description{
Adds learners from the {survival} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

}
